Chore : 사용하지 않는 주석 제거
